% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.code.r
\docType{package}
\name{chatham.bio532-package}
\alias{chatham.bio532}
\alias{chatham.bio532-package}
\title{Statistical analyses especially designed for BIO532, Biostatistics, at Chatham University}
\description{
Functions in this package allow one to employ the panoply of common statistical analyses, 
but with a twist.  Every inferential test also uses a resampling procedure to generate empirical probability 
distributions for null hypotheses.  Simulated and theoretical probability distributions can be compared, 
diagnostics for tests are available, and results/output are presented in a logical style, consistent with 
developing appropriate interpretations.  This package is designed to help BIO532 students navigate the BIO532
course, but it should also have general appeal.
}
\section{BIO532 TOC}{

BIO532-package
}
\author{
Michael Collyer
}

