% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.int.r
\name{confidence.int}
\alias{confidence.int}
\title{Generate confidence intervals}
\usage{
confidence.int(y, alpha = 0.05, iter = 1000)
}
\arguments{
\item{y}{A vector of data}

\item{alpha}{The significance level}

\item{iter}{Number of resampling (bootstrap) iterations}
}
\value{
The function returns a list of samples, means from samples, and (1-alpha)*100%
confidence intervals.
}
\description{
Function generates one-sample confidence intervals from a resampling (bootstrap) procedure
}
\examples{

Y <- rnorm(10000) # population
y <- sample(Y, size = 30) # sample from population
CI <- confidence.int(y, alpha = 0.05, iter=1000)
summary(CI)
plot(CI)
mean(Y) # does confidence interval contain population mean?
}
\author{
Michael Collyer
}

