% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisample.r
\name{multisample}
\alias{multisample}
\title{Generate multiple samples}
\usage{
multisample(population, size, permutations = 1000, CLT = FALSE)
}
\arguments{
\item{population}{A vector of data for the whole population}

\item{size}{The size of the sample}

\item{permutations}{The number of times to sample from the population}

\item{CLT}{A logical value to indicate if summary statistics should be generated, 
# based on the central limit theorem.}
}
\value{
The function returns a list of samples.  If CLT = TRUE, it also returns summary
statistics from the samples, plus expected values based on the central limit theorem.
}
\description{
Function samples many times from the same population.
}
\examples{

Y <- rnorm(10000)
mySamples <- multisample(population = Y, size = 50, permutations = 100, CLT = FALSE)
summary(mySamples)
mySamples$samples[1:3]

mySamples <- multisample(population = Y, size = 50, permutations = 100, CLT = TRUE)
summary(mySamples)
plot(mySamples)

}
\author{
Michael Collyer
}

