% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.means.test.R
\name{pairwise.means.test}
\alias{pairwise.means.test}
\title{Perform non-parametric (randomization) test of mean differences for ANOVA}
\usage{
pairwise.means.test(A, alpha = 0.05)
}
\arguments{
\item{A}{A class ANOVA object from a previous analysis}

\item{alpha}{The significance level for confidence intervals.}
}
\value{
The function returns a table of confidence intervals for all pairwise comparisons
of means.
}
\description{
Function performs all possible pairwise tests of mean differences for a class ANOVA
object, by calculating confidence intervals
}
\examples{
data(cad)
cad$center <- as.factor(cad$center) # change numeric variable to factor

ANOVA1 <- ANOVA(age ~ center, data = cad, iter = 999)
summary(ANOVA1)
plot(ANOVA1, method = "hist")
pw <- pairwise.means.test(ANOVA1)
summary(pw)
plot(pw)
}
\author{
Michael Collyer
}

