% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{plot.ANOVA}
\alias{plot.ANOVA}
\title{Plot Function for chatham.bio532}
\usage{
\method{plot}{ANOVA}(x, method = c("histogram", "diagnostic"), ...)
}
\arguments{
\item{x}{plot object (from \code{\link{ANOVA}})}

\item{method}{Whether to plot histograms of test statistic distributions or linear model diagnostics}

\item{...}{other arguments passed to plot (see hist)}
}
\description{
Plot Function for chatham.bio532
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

