% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{plot.confidence.int}
\alias{plot.confidence.int}
\title{Plot Function for chatham.bio532}
\usage{
\method{plot}{confidence.int}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{confidence.int}})}

\item{...}{other arguments passed to plot (see hist)}
}
\description{
Plot Function for chatham.bio532
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

