% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{plot.multisample}
\alias{plot.multisample}
\title{Plot Function for chatham.bio532}
\usage{
\method{plot}{multisample}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{multisample}})}

\item{...}{other arguments passed to plotAllSpecimens (see hist)}
}
\description{
Plot Function for chatham.bio532
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

