% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{plot.pairwise.means.test}
\alias{plot.pairwise.means.test}
\title{Plot Function for chatham.bio532}
\usage{
\method{plot}{pairwise.means.test}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{pairwise.means.test}})}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for chatham.bio532
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

