% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{plot.two.sample.test}
\alias{plot.two.sample.test}
\title{Plot Function for chatham.bio532}
\usage{
\method{plot}{two.sample.test}(x, method = c("histogram", "diagnostic"),
  conf.int = c("2T", "PT", "NT"), ...)
}
\arguments{
\item{x}{plot object (from \code{\link{two.sample.test}})}

\item{method}{Whether to plot histograms of test statistic distributions or linear model diagnostics}

\item{conf.int}{If histogram is chosen for method, the choice for confidence interval display.  2T produces a 
two-tailed interpretation of a confidence interval; PT produces a positive limit (if the alternative
hypothesis is expected to be "greater than"); NT produces a negative limit (if the alternative
hypothesis is expected to be "less than").  The confidence intervals are automatically found from the alpha 
argument of the test that was performed.}

\item{...}{other arguments passed to plot (see hist)}
}
\description{
Plot Function for chatham.bio532
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

