% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{print.multisample}
\alias{print.multisample}
\title{Print/Summary Function for chantahm.bio532}
\usage{
\method{print}{multisample}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{gpagen}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for chantahm.bio532
}
\author{
Michael Collyer
}
\keyword{utilities}

