% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{summary.ANOVA}
\alias{summary.ANOVA}
\title{Print/Summary Function for chatham.bio532}
\usage{
\method{summary}{ANOVA}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{ANOVA}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for chatham.bio532
}
\author{
Michael Collyer
}
\keyword{utilities}

