% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.code.r
\docType{data}
\name{twenty}
\alias{twenty}
\title{twenty data from Pagano and Gauvreau}
\description{
twenty data from Pagano and Gauvreau
}
\references{
Principles of Biostatistics, 2nd ed., 
by Marcello Pagano and Kimberlee Gauvreau (2000) 
Pacific Grove, PA: Duxbury publishers. ISBN: 0-534-22902-6 
U.S.A. 97:4106-4111
}
\keyword{datasets}

